[cudamapper] allocator refactor
